+++ price-adapter-raw/src/lib.rs
@@ -64,8 +64,10 @@
 
 mod binance_websocket;
 mod coingecko;
+mod coinmarketcap;
 pub mod error;
 pub mod types;
 
 pub use binance_websocket::{BinanceWebsocket, BinanceWebsocketService};
 pub use coingecko::CoinGecko;
+pub use coinmarketcap::CoinMarketCap;