+++ price-adapter/src/mappers/band_static_mapper.rs
@@ -28,6 +28,10 @@ impl BandStaticMapper {
                 env!("CARGO_MANIFEST_DIR"),
                 "/resources/coingecko.json"
             ))),
+            "coinmarketcap" => Ok(include_str!(concat!(
+                env!("CARGO_MANIFEST_DIR"),
+                "/resources/coinmarketcap.json"
+            ))),
             _ => Err(Error::UnsupportedSource), // Add more matches for other sources
         }?;